+++ web/src/AacLibraryPage.test.tsx
@@ -22,6 +22,7 @@ const adminAccount: AccountPublic = {
   organizationId: 'org-1',
   name: null,
   emailVerified: true,
+  mustChangePassword: false,
 };
 
 function makeSymbol(
@@ -77,6 +78,7 @@ function fakeApi(): Api {
     updateSettings: notImplemented,
     deleteUser: () => Promise.resolve(),
     createCaregiverAccount: notImplemented,
+    listAccounts: notImplemented,
     listCaregivers: notImplemented,
     linkCaregiver: notImplemented,
     generateDeviceCode: notImplemented,