+++ web/src/PersonalContextPanel.test.tsx
@@ -72,6 +72,7 @@ function fakeApi(): Api {
     updateSettings: notImplemented,
     deleteUser: notImplemented,
     createCaregiverAccount: notImplemented,
+    listAccounts: notImplemented,
     listCaregivers: notImplemented,
     linkCaregiver: notImplemented,
     generateDeviceCode: notImplemented,